JUS-108 Reorganizing files and changing how user information is retrieved
