+++ client/src/index.js
@@ -8,7 +8,7 @@ import OrgVerifyEmail from './pages/user/organizer/verifyEmail/verifyEmail.jsx'
 import OrgVerificationRequired from './pages/user/organizer/verificationRequired/verificationRequired.jsx'
 import CustomerVerifyEmail from './pages/user/verifyEmail/verifyEmail.jsx'
 import CustomerVerificationRequired from './pages/user/verificationRequired/verificationRequired.jsx'
-import CustomerInfo from './pages/user/customer/customerInfo/customerInfo.jsx'
+import CustomerInfo from './pages/user/customerInfo/customerInfo.jsx'
 import Header from './components/header/Header'
 import { store } from './store.js'
 import {Provider} from 'react-redux'
